Move transactionPrice calculation to hook from component
